pyproject.toml
[tool.poetry]
name = "\\"
version = "0.1.0"
description = "chatPDF project"
authors = ["arnaldocavalcanti <naldocavalcanti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
langchain = "^0.2.5"
openai = "^1.34.0"
langchain-community = "^0.2.5"
pypdf2 = "^3.0.1"
streamlit-chat = "^0.1.1"
tiktoken = "^0.7.0"
faiss-cpu = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
